move attributeFunctions to utilties
